refactor(core)!: move StakePool.metadata under PoolParameters
